# Add: exponential backoff for CAS operations on floats