@imorph 2024-10-26
`TestSummaryDecay` test is failing, but this fail seems unrelated to my changes (may be I missing something but looks like it is different code path).

Most likely it fils because in the `summary` struct, observations added via `Observe` are buffered in `hotBuf`. When `hotBuf` reaches its capacity or expires, an asynchronous flush is triggered by calling `asyncFlush`. This flush moves the observations to `coldBuf` and processes them in a separate goroutine. The `Write` method processes the `coldBuf` synchronously but does not include observations still in `hotBuf` or those being processed asynchronously in the background goroutine. Test is rapidly adding observations and periodically calling `sum.Write()`. Due to the asynchronous processing of observations, some observations may not have been fully processed and included in the quantile calculations when `Write` is called. This could lead to random test failures.

By i might be wrong.