@imorph 2024-11-05
@ArthurSens 

> This is only called if the first CAS fails, which doesn't happen on the happy path. 

It is only happy in single-threaded case, as long as we have contending goroutines trying to CAS atomic value, inevitably we have a lot of sleeps which produce (still unsure =) ) syscalls and it is not free. Some projects sensitive to increasing number of syscalls, latest example from same folks at cockroachdb: https://github.com/cockroachdb/cockroach/issues/133315 
 
> Or the number of syscalls that are potentially increasing?

yes, exactly that