@ArthurSens 2024-11-05
[Golang's sleep implementation surely does syscalls](https://github.com/golang/go/issues/25471), how much of a problem that is uncertain to me.

[One suggestion](https://github.com/cockroachdb/cockroach/issues/133315#issuecomment-2434294919) from the link you provided is to use `time.Since` or maybe you could use a `time.Ticker` if we could find a way to make it exponential. What do you think? What do you think?