@bwplotka 2024-11-11
Epic work @ArthurSens on checking alternatives, and @imorph for trying those out ❤️ 

There is also https://github.com/prometheus/client_golang/issues/1642 that discusses more complex algorithms for concurrency (adaptive locks), we can continue the discussion there.