@ywwg 2024-11-19
I believe adding a sleep to the TestHistogramAtomicObserve test in the observe() func would resolve the issue.  @bwplotka can you take a look?